Moved Lottery.java class to ball.lottery package and added necessary imports
